# Improve countlines() performance