Size of File | Base.countlines | new countlines
---- | ---- | -------
10MB | 11.3 milliseconds | 7.3 milliseconds
100MB | 102 milliseconds | 62.5 milliseconds
1GB | 928 milliseconds | 598 milliseconds
21GB | 31 seconds | 26 seconds

The first 3 files were auto-generated from random characters + newlines. The 21GB file is from @jiahao's test file from the readdlm issue.

Also added previously missing tests