@carnaval 2015-06-30
Is it computing the same thing ? The old impl. seems to not be counting empty lines.