@quinnj 2015-06-30
Yeah, I noticed that. I'm not sure why we wouldn't count empty lines? Looking at the source code for `wc` it didn't have anything like that.