@quinnj 2015-06-30
We could maybe provide a `skip_emtpy` argument, but I'd really like to know if anyone had a use for this.