@mbauman 2015-06-30
Well, it was introduced in https://github.com/JuliaLang/julia/commit/2eb249331bf0da3511f979d86e2c0bcb36e05be5 with a reference to dlmread.  I didn't dig any further, but is it (or was it) used to determine the size of an array to preallocate?