@ScottPJones 2015-06-30
OK, I've looked everywhere.  This is documented as counting non-empty lines, however, it is not used anywhere in Base, and is only used in one place in any registered package, in `Jumos/src/Trajectories/XYZ/Reader.jl`, where it is just verifying that the number of lines stored in the header of the file matches the rest of the lines in the file, not performance critical.