@quinnj 2015-06-30
Thanks for looking around @ScottPJones. My guess is it used to be used by `readdlm`, but now it's using something else. Sounds like the breaking nature here won't affect much, so I'm inclined to merge and we can add another argument later if a compelling use case comes up. I'll update the docs.