@ScottPJones 2015-06-30
Yes, I'm sad though, because your code looks very nice and fast, only to find out that it's not (currently) used except that once in a package.
Better to merge it in, have the faster,better code for anybody in the future who needs it!