@quinnj 2015-06-30
Ref: https://github.com/Luthaf/Jumos.jl/issues/22